+++ Window.cpp
@@ -63,10 +63,10 @@ void Window::change_symbol(std::string symbol, bool state) {
 	screen_symbols[std::find(Graphics::ScreenSymbols.begin(), Graphics::ScreenSymbols.end(), symbol) - Graphics::ScreenSymbols.begin()] = state;
 }
 
-void Window::handle_key_down(int key) {
+void Window::handle_key_down(uint8_t key) {
 
 }
 
-void Window::handle_key_up(int key) {
+void Window::handle_key_up(uint8_t key) {
 
 }
